requirements.txt
python-telegram-bot==20.7
yfinance
matplotlib
nest_asyncio